feat(robots): extract renderRobots() into @takazudo/zudo-doc/robots (#2324)
Ship renderRobots(settings) from packages/zudo-doc/src/robots.ts with a
new ./robots exports-map entry. Collapse both pages/robots.txt.tsx and
the create-zudo-doc template counterpart to ~4-line stubs that delegate
to the package helper, eliminating copy-drift between showcase and scaffold.
